Renamed detector classes to calodetector
